requirements.txt
langchain
groq
fastapi
uvicorn
python-dotenv
textblob
sqlite3